requirements.txt
absl-py==0.13.0
astor==0.8.1
bokeh==1.4.0
cached-property==1.5.2
gast==0.2.2
google-pasta==0.2.0
grpcio==1.39.0
h5py==2.10.0
importlib-metadata==4.6.1
Jinja2==3.0.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.3.4
MarkupSafe==2.0.1
numpy==1.21.1
opt-einsum==3.3.0
packaging==21.0
Pillow==8.3.1
protobuf==3.17.3
pyparsing==2.4.7
python-dateutil==2.8.2
PyYAML==5.4.1
six==1.16.0
tensorboard==1.15.0
tensorflow-cpu==1.15.0
tensorflow-cpu-estimator==1.15.1
termcolor==1.1.0
tornado==6.1
typing-extensions==3.10.0.0
Werkzeug==2.0.1
wrapt==1.12.1
zipp==3.5.0
